feat: complete Messaging Module refactoring
- Move ILoggingConfigurationService and LoggingConfigurationService to Messaging module
- Update namespaces to ClubDoorman.Services.Messaging
- Update checklist progress to 30% (6/20 phases completed)
- All tests pass successfully
